Address PR 29 review follow-ups
Resolve unresolved review feedback by expanding context-key tests,
fixing numeric NaN truthiness semantics, splitting tokenizer logic
out of context-key-parser.ts to satisfy the 400-line file limit,
and applying requested execplan wording fixes.

Kept parser behaviour and syntax diagnostics stable while preserving
the SourceReader/Position model requested in this milestone.
